ride-matching-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
pydantic==2.4.2
grpcio==1.59.3
grpcio-tools==1.59.3
